Specifying Sample Spaces of Graphs

The sample space for a model defines the set of all graphs our model assigns probability to. For example, we may wish to only consider directed graphs with binary edges and no self loops, which corresponds to graphs of type spaces.SparseDirectedGraph. This particular type also stores graphs as sparse arrays and it suited for large graphs with low edge density.

Matrix Representation

To keep things consistent, adjacency matrices of graphs will use the i,j entry to indicate the presence of edge i -> j.

Pre-defined Sample Spaces

ergm.spaces.SparseDirectedGraph — Type

Directed graphs backed by sparse matrices.

This type includes a parameter n to specify the number of nodes. These graphs also have no self loops and edges are binary. To create an empty graph on n nodes, use SparseDirectedGraph{n}(). To create a graph from a sparse adjacency matrix adjacency, use SparseDirectedGraph(adjacency).

Edge indices for this type take the form (s, t) where s::Int and t::Int are the indices of the respective source and target nodes of the edge. Node indices are integers i in the range 1 ≤ i ≤ n, where n is the type paramter specifying the number of nodes in the graph. For example, querying the value of the edge from node 1 to node 3 in the graph x :: SparseDirectedGraph is accomplished by getindex(x, (1, 3)) or, equivalently, x[(1, 3)].

Custom Sample Spaces

A space of graphs is a subtype of spaces.SampleSpace that implements the below interface. In addition to the below interface, all sample spaces must have a constructor that takes no arguments are returns a value corresponding to an empty graph. This is used to initialize samplers, for example. Currently, it is assumed that edges are binary in all SampleSpaces and this interface will need to be augmented to support more edge values.

ergm.spaces.random_index — Method

random_index(g::SampleSpace)

Sample the index of an edge uniformly. The edge need not actually be present.

Base.getindex — Method

getindex(g::SampleSpace, index)

Retrieve value of edge with given index.

Base.setindex! — Method

setindex!(g::SampleSpace, value, index)

Set value of edge with given index.

Base.copy — Method

copy(g::SampleSpace)

Perform a deep copy of an element of the space.