Sampling from ERGMs

Sampling from an ERGM is accomplished by constructing a sampling.GibbsSampler object from a chosen ERGM (an object of type models.Model) and passing it to the function sampling.sample.

ergm.sampling.GibbsSampler — Type

GibbsSampler(model::Model; burn_in::Int, sample_interval::Int)

Construct a Gibbs sampler for an given ERGM.

A Gibbs sampler implements a Markov chain where each step can toggle at most one edge in the state graph. Before returning the first sample, burn_in steps are taken along the Markov chain. After the first sample, sample_interval steps are taken between each pair of samples returned with the goal of producing approximately independent samples.

Choosing Sampler Parameters

ergm.sampling.plot_diagnostics — Function

Plot diagnostic functions useful for choosing Gibbs sampler parameters.

This function takes steps steps along the Gibbs sampler Markov chain starting from an empty graph and records the value of the graph statistics at every individual step. This is used to compute rolling means of each statistics with window size window_size as well as the autocorrelation function of each of the statistics.

Typically, we want to choose a burn-in length for the sampler long enough for all the rolling means of the statistics to stabilize, which can be judged heuristically just by looking at the plot of the rolling means.

We also need to choose an interval between samples long enough that the autocorrelation of each of the statistics drops off near zero between returned samples, which is a necessary condition for the sampler to generate high-equality, approximately independent samples.

This function returns a matrix of all graph statistics, a matrix of graph statistics smoothed with the given window size, a matrix of autocorrelation functions, and the plot object.